+++ services/storage/src/simcore_service_storage/modules/db/__init__.py
@@ -1,8 +1,7 @@
 import logging
 
 from fastapi import FastAPI
-from servicelib.db_async_engine import close_db_connection
-from servicelib.fastapi.db_asyncpg_engine import connect_to_db
+from servicelib.fastapi.db_asyncpg_engine import close_db_connection, connect_to_db
 from servicelib.retry_policies import PostgresRetryPolicyUponInitialization
 from sqlalchemy.ext.asyncio import AsyncEngine
 from tenacity import retry
